backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
python-dotenv==1.0.1
marshmallow==3.21.3
SQLAlchemy==2.0.31
google-generativeai==0.7.2
Werkzeug==3.0.3